Refactor ChumBucketHider to be in fishing
