+++ src/main/java/at/hannibal2/skyhanni/SkyHanniMod.kt
@@ -164,7 +164,7 @@ import at.hannibal2.skyhanni.features.misc.BrewingStandOverlay
 import at.hannibal2.skyhanni.features.misc.ButtonOnPause
 import at.hannibal2.skyhanni.features.misc.ChestValue
 import at.hannibal2.skyhanni.features.misc.ChickenHeadTimer
-import at.hannibal2.skyhanni.features.misc.ChumBucketHider
+import at.hannibal2.skyhanni.features.fishing.ChumBucketHider
 import at.hannibal2.skyhanni.features.misc.CollectionTracker
 import at.hannibal2.skyhanni.features.misc.CompactSplashPotionMessage
 import at.hannibal2.skyhanni.features.misc.CurrentPetDisplay
